idSearch, pwSearch, purchaseList, unMemberPurchaseList JavaScript작성

+++ src/main/java/com/book/book_project/dto/FavoritesDTO.java
@@ -16,7 +16,7 @@ import java.sql.Timestamp;
 @NoArgsConstructor
 @AllArgsConstructor
 public class FavoritesDTO {
-    private Long favoriteseq;
+    private int favoriteseq;
     private Timestamp favoritedate;
     private String favoritecheck;
     private MemberEntity userid;
